# openapi/queue-management-api.yaml
openapi: 3.1.0
info:
  title: Queue Management API
  version: 0.1.0
  description: API-first contract for the Queue Management App MVP.
servers:
  - url: http://localhost:5000
    description: Local development
paths:
  /api/health:
    get:
      summary: Check API health
      operationId: getHealth
      tags:
        - Health
      responses:
        '200':
          description: API is running.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: Healthy
  /api/owners/register:
    post:
      summary: Register an owner and first queue location
      operationId: registerOwner
      tags:
        - Owners
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerRegistrationRequest'
            example:
              ownerName: Priya Sharma
              email: priya@example.com
              mobile: '9876543210'
              password: StrongPass123
              businessName: Priya Dental Clinic
              locationName: Main Branch
              address: 12 MG Road, Bengaluru
              businessMobile: '9876500000'
      responses:
        '201':
          description: Owner account and queue location were created.
          headers:
            Location:
              description: Public queue location URL.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerRegistrationResponse'
              example:
                ownerId: 1
                queueLocationId: 1
                locationCode: AB7K2M9Q
                businessName: Priya Dental Clinic
                locationName: Main Branch
                role: Owner
        '400':
          description: Registration request is invalid.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
        '409':
          description: Owner email or mobile number is already registered.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/auth/login:
    post:
      summary: Log in as an owner or manager
      operationId: login
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              email:
                value:
                  identifier: priya@example.com
                  password: StrongPass123
              mobile:
                value:
                  identifier: '9876543210'
                  password: StrongPass123
      responses:
        '200':
          description: Login succeeded and returned a bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                tokenType: Bearer
                expiresIn: 3600
                userId: 1
                name: Priya Sharma
                roles:
                  - Owner
        '400':
          description: Login request is invalid.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
        '401':
          description: Identifier or password is invalid.
  /api/auth/me:
    get:
      summary: Get the current authenticated user
      operationId: getCurrentUser
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current authenticated user details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '401':
          description: Missing or invalid bearer token.
  /api/locations/{locationCode}/queue-entries:
    post:
      summary: Join a public customer queue
      operationId: joinCustomerQueue
      tags:
        - Customer Queue
      parameters:
        - name: locationCode
          in: path
          required: true
          schema:
            type: string
          example: AB7K2M9Q
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerJoinQueueRequest'
            example:
              customerName: Amit Kumar
              mobile: '9876543210'
              partySize: 2
              serviceReason: Consultation
      responses:
        '201':
          description: Customer joined the queue and received a private status link.
          headers:
            Location:
              description: Private customer status URL.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerJoinQueueResponse'
              example:
                queueEntryId: 10
                queueLocationId: 1
                locationCode: AB7K2M9Q
                tokenNumber: 1
                status: Waiting
                trackingToken: rA7kXJ3mR1T8vF2pQ9sL0wY6nB4cD5eG
                statusUrl: /status/10/rA7kXJ3mR1T8vF2pQ9sL0wY6nB4cD5eG
        '400':
          description: Join request is invalid.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
        '404':
          description: Location code was not found.
        '409':
          description: Queue is currently closed.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/manager/queue/status:
    get:
      summary: Get the authenticated manager queue status
      operationId: getManagerQueueStatus
      tags:
        - Manager Queue
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current queue status for the user's assigned location.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagerQueueStatusResponse'
        '401':
          description: Missing or invalid bearer token.
        '404':
          description: User does not have an assigned queue location.
  /api/manager/queue/open:
    post:
      summary: Open the authenticated manager queue
      operationId: openManagerQueue
      tags:
        - Manager Queue
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Queue was opened.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagerQueueStatusResponse'
        '401':
          description: Missing or invalid bearer token.
        '404':
          description: User does not have an assigned queue location.
  /api/manager/queue/close:
    post:
      summary: Close the authenticated manager queue
      operationId: closeManagerQueue
      tags:
        - Manager Queue
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Queue was closed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagerQueueStatusResponse'
        '401':
          description: Missing or invalid bearer token.
        '404':
          description: User does not have an assigned queue location.
  /api/manager/queue/reset:
    post:
      summary: Reset today's queue for the authenticated manager
      operationId: resetManagerQueue
      tags:
        - Manager Queue
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Today's queue entries were cleared and the queue was closed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagerQueueStatusResponse'
        '401':
          description: Missing or invalid bearer token.
        '404':
          description: User does not have an assigned queue location.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: Healthy
    OwnerRegistrationRequest:
      type: object
      required:
        - ownerName
        - password
        - businessName
        - address
        - businessMobile
      properties:
        ownerName:
          type: string
          minLength: 1
          maxLength: 100
          example: Priya Sharma
        email:
          type:
            - string
            - 'null'
          format: email
          maxLength: 254
          example: priya@example.com
        mobile:
          type:
            - string
            - 'null'
          maxLength: 30
          example: '9876543210'
        password:
          type: string
          minLength: 8
          maxLength: 100
          format: password
          example: StrongPass123
        businessName:
          type: string
          minLength: 1
          maxLength: 150
          example: Priya Dental Clinic
        locationName:
          type:
            - string
            - 'null'
          maxLength: 150
          example: Main Branch
        address:
          type: string
          minLength: 1
          maxLength: 300
          example: 12 MG Road, Bengaluru
        businessMobile:
          type: string
          minLength: 1
          maxLength: 30
          example: '9876500000'
      anyOf:
        - required:
            - email
        - required:
            - mobile
    OwnerRegistrationResponse:
      type: object
      required:
        - ownerId
        - queueLocationId
        - locationCode
        - businessName
        - role
      properties:
        ownerId:
          type: integer
          format: int32
          example: 1
        queueLocationId:
          type: integer
          format: int32
          example: 1
        locationCode:
          type: string
          example: AB7K2M9Q
        businessName:
          type: string
          example: Priya Dental Clinic
        locationName:
          type:
            - string
            - 'null'
          example: Main Branch
        role:
          type: string
          enum:
            - Owner
          example: Owner
    LoginRequest:
      type: object
      required:
        - identifier
        - password
      properties:
        identifier:
          type: string
          minLength: 1
          maxLength: 254
          example: priya@example.com
        password:
          type: string
          minLength: 1
          maxLength: 100
          format: password
          example: StrongPass123
    LoginResponse:
      type: object
      required:
        - accessToken
        - tokenType
        - expiresIn
        - userId
        - name
        - roles
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          enum:
            - Bearer
        expiresIn:
          type: integer
          format: int32
          example: 3600
        userId:
          type: integer
          format: int32
          example: 1
        name:
          type: string
          example: Priya Sharma
        roles:
          type: array
          items:
            type: string
            enum:
              - Owner
              - Manager
    CurrentUserResponse:
      type: object
      required:
        - userId
        - name
        - roles
      properties:
        userId:
          type: integer
          format: int32
        name:
          type: string
        roles:
          type: array
          items:
            type: string
    CustomerJoinQueueRequest:
      type: object
      required:
        - customerName
      properties:
        customerName:
          type: string
          minLength: 1
          maxLength: 100
          example: Amit Kumar
        mobile:
          type:
            - string
            - 'null'
          maxLength: 30
          example: '9876543210'
        partySize:
          type:
            - integer
            - 'null'
          format: int32
          minimum: 1
          maximum: 50
          example: 2
        serviceReason:
          type:
            - string
            - 'null'
          maxLength: 200
          example: Consultation
    CustomerJoinQueueResponse:
      type: object
      required:
        - queueEntryId
        - queueLocationId
        - locationCode
        - tokenNumber
        - status
        - trackingToken
        - statusUrl
      properties:
        queueEntryId:
          type: integer
          format: int32
          example: 10
        queueLocationId:
          type: integer
          format: int32
          example: 1
        locationCode:
          type: string
          example: AB7K2M9Q
        tokenNumber:
          type: integer
          format: int32
          example: 1
        status:
          type: string
          enum:
            - Waiting
          example: Waiting
        trackingToken:
          type: string
        statusUrl:
          type: string
          example: /status/10/rA7kXJ3mR1T8vF2pQ9sL0wY6nB4cD5eG
    ManagerQueueStatusResponse:
      type: object
      required:
        - queueLocationId
        - locationCode
        - businessName
        - isQueueOpen
        - waitingCount
      properties:
        queueLocationId:
          type: integer
          format: int32
          example: 1
        locationCode:
          type: string
          example: AB7K2M9Q
        businessName:
          type: string
          example: Priya Dental Clinic
        isQueueOpen:
          type: boolean
          example: true
        waitingCount:
          type: integer
          format: int32
          example: 0
        currentTokenNumber:
          type:
            - integer
            - 'null'
          format: int32
          example: 1
    ValidationProblemDetails:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string